Fix possible logic issue
This patch addresses both SC2155 and SC2181 which may have caused incorrect
error checking of the `git rev-parse` call.

it was then discovered that GetOid() may be returning the wrong value.
It appears that GetOid() should be returning commitOid and not oid as it
currently does.

This patch addresses all three problems.

In ditto line 92:
        local oid=$(git rev-parse HEAD)
              ^-^ SC2155: Declare and assign separately to avoid masking return values.
In ditto line 94:
        if [ "$?" -ne "0" ]; then
             ^--^ SC2181: Check exit code directly with e.g. 'if mycmd;', not indirectly with $?.